fix: fix knowledge panel content building
We had Crowdin issue that led to quality issue, as any key was
translatable, and as Crowdin uses the position of the item in the YAML
file as a key (which obviously does not work well with reorganizing
the items).

I simplify here the YAML file structure: it's now only a dict, with
the key being the value tag, and the value being a dict with currently
only the "content" field.

This should work well with Crowdin.