feat: fix knowledge panel outputs (#754)
- `knowledge_panels/labels/off/world-en.yml` had an incorrect structure:
the YAML should be a mapping with the key being the tag. I fixed the
structure for the reference file `world-en.yml`, and deleted the
translation files.
- I removed the previous knowledge panel directory structure
(`lang/{lang}/knowledge_panels/{tag_type}`) to only keep the new
structure with the flavor
(`lang/{flavor}/{lang}/knowledge_panels/{tag_type}`)
- don't generate an HTML file if the content is the same as the
reference language (en). This greatly reduces the number of files, and
prevent excessive memory usage from Product Opener (there is an English
fallback if the translation is not available on Product Opener anyway).