fix: broken text center alignment on page 2 of pro platform (#781)
This PR fixes
https://github.com/openfoodfacts/openfoodfacts-server/issues/12228

The issue is a `<center>` tag that was added around the `<h1>` tag. It
is kept as-is on page 1, but for other pages, we don't want to display
what's after the title, so all the comment after `</h1>` is removed.
That means we have a `<center>` but no `</center>`.

I changed the fr and en pages, but the other translations should be
resynched too...

---------

Co-authored-by: OFF <off@openfoodfacts.org>