docs: typo fixup: `ot` -> `of` (#803)
### What
Typo fixup for the `terms-of-service.html` page in each affected
language.

I spotted this while preparing to print copies of the ODBL and Database
Contents License files to bring with me for OpenFoodFacts Days 2025
(because I brought printed copies of products from the database with me
for a workshop).

### Screenshot
N/A

### Fixes bug(s)
- Resolves #794.

### Part of 
N/A