fix: use presskit for /press in French? (#839)
This PR is to serve presskit.html when users access
https://world.openfoodfacts.org/presse

It's not very clear to me which press page we want to serve at which
URL.

On the website, the link to the press page is /press (or /presse in
French).

There used to be a nginx redirect (now broken) to use presskit when
press was accessed.

In openfoodfacts-web, we have different content for press.html and
presskit.html. There are also sometimes local version like presse.html
that are different too.

My recommendation:

- decide if we want to have 2 different pages, one for /press and one
for /presskit (what would be the difference?)
- if there's only one page, decide which url it should be. I think
/press is easier (and /presse etc. in French and other languages) and
avoids issues with remembering if it's /presskit or /press-kit
- move the content we want to the corresponding page: e.g. press.html
(and remove presskit.html and local versions like presse.html)

Co-authored-by: Off <off@openfoodfacts.org>